requirements.txt
aiofiles==25.1.0
aiogram==3.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
asyncpg==0.31.0
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Django==6.0.1
docxtpl==0.20.2
fastapi==0.127.1
frozenlist==1.8.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
lxml==6.0.2
magic-filter==1.0.12
MarkupSafe==3.0.3
multidict==6.7.0
pillow==12.1.0
propcache==0.4.1
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-docx==1.2.0
python-dotenv==1.2.1
python-telegram-bot==22.5
requests==2.32.5
soupsieve==2.8.3
SQLAlchemy==2.0.45
sqlparse==0.5.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
yarl==1.22.0
